refactor(tools): use #7262's strictUrlPathSegment and delete the local copy
#7262 landed d2c74d7416, defining strictUrlPathSegment and
strictEncodedUrlPathSegment in url-path.ts to refuse padded identifiers on
state-changing requests. That is the rule this branch introduced locally while
the two PRs were in flight, so the duplication collapses now that the rebase
brings it in: tools/strict-url-path.ts is deleted and its four consumers import
from @/tools/url-path.

Their assertUnpadded is slightly better than the local version — an all-whitespace
value falls through to safeUrlPathSegment and reports "is required" rather than
a padding error, which names the real problem. strictCanonicalBigQueryId now
derives from their guard too, so the body value and the path value share one
rule rather than two.

The error text changed from "cannot have" to "must not have leading or trailing
whitespace", and four assertions failed on the rebase because they pin the exact
text. They are updated to the new wording rather than loosened — that precision
is the property that caught this and two earlier upstream changes.
