+++ apps/sim/tools/box_sign/cancel_request.ts
@@ -1,5 +1,5 @@
-import { strictUrlPathSegment } from '@/tools/strict-url-path'
 import type { ToolConfig } from '@/tools/types'
+import { strictUrlPathSegment } from '@/tools/url-path'
 import type { BoxSignCancelRequestParams, BoxSignResponse } from './types'
 import { SIGN_REQUEST_OUTPUT_PROPERTIES } from './types'
